User Interface(fix): relocating entities within model & changes IUser into IUserRegisterRequest
